worker: log reason if the workflow errors
