+++ packages/ws-worker/src/events/atttempt-complete.ts
@@ -1,9 +1,9 @@
-import { timestamp } from '@openfn/logger';
 import type { WorkflowCompletePayload } from '@openfn/engine-multi';
 
 import { ATTEMPT_COMPLETE, AttemptCompletePayload } from '../events';
 import { calculateAttemptExitReason } from '../api/reasons';
-import { sendEvent, Context, onJobLog } from '../api/execute';
+import { sendEvent, Context } from '../api/execute';
+import logFinalReason from '../util/log-final-reason';
 
 export default async function onWorkflowComplete(
   context: Context,
@@ -14,21 +14,9 @@ export default async function onWorkflowComplete(
   // TODO I dont think the attempt final dataclip IS the last job dataclip
   // Especially not in parallelisation
   const result = state.dataclips[state.lastDataclipId!];
+
   const reason = calculateAttemptExitReason(state);
-
-  const time = (timestamp() - BigInt(10e6)).toString();
-
-  let message = `Run complete with status: ${reason.reason}`;
-  if (reason.reason !== 'success') {
-    message += `\n${reason.error_type}: ${reason.error_message || 'unknown'}`;
-  }
-
-  await onJobLog(context, {
-    time,
-    message: [message],
-    level: 'info',
-    name: 'R/T',
-  });
+  await logFinalReason(context, reason);
 
   await sendEvent<AttemptCompletePayload>(channel, ATTEMPT_COMPLETE, {
     final_dataclip_id: state.lastDataclipId!,
