+++ packages/ws-worker/test/api/execute.test.ts
@@ -14,7 +14,6 @@ import {
   onJobLog,
   execute,
   onWorkflowStart,
-  onWorkflowError,
   loadDataclip,
   loadCredential,
   sendEvent,
@@ -242,113 +241,6 @@ test('workflowStart should send an empty attempt:start event', async (t) => {
 
 //   await onWorkflowComplete(context, event);
 // });
-
-test('workflowError should trigger runComplete with a reason', async (t) => {
-  const jobId = 'job-1';
-
-  const state = {
-    reasons: {},
-    dataclips: {},
-    lastDataclipId: 'x',
-    activeJob: jobId,
-    activeRun: 'b',
-    errors: {},
-  };
-
-  const channel = mockChannel({
-    [RUN_COMPLETE]: (evt) => {
-      t.is(evt.reason, 'crash');
-      t.is(evt.error_message, 'it crashed');
-      return true;
-    },
-    [ATTEMPT_COMPLETE]: () => true,
-  });
-
-  const event = {
-    severity: 'crash',
-    type: 'Err',
-    message: 'it crashed',
-  };
-
-  const context = { channel, state, onFinish: () => {} };
-
-  await onWorkflowError(context, event);
-});
-
-test('workflow error should send reason to onFinish', async (t) => {
-  const jobId = 'job-1';
-
-  const state = {
-    reasons: {},
-    dataclips: {},
-    lastDataclipId: 'x',
-    activeJob: jobId,
-    activeRun: 'b',
-    errors: {},
-  };
-
-  const channel = mockChannel({
-    [RUN_COMPLETE]: (evt) => true,
-    [ATTEMPT_COMPLETE]: () => true,
-  });
-
-  const event = {
-    error: {
-      severity: 'crash',
-      type: 'Err',
-      message: 'it crashed',
-    },
-    state: {},
-  };
-
-  const context = {
-    channel,
-    state,
-    onFinish: (evt) => {
-      t.is(evt.reason.reason, 'crash');
-    },
-  };
-
-  await onWorkflowError(context, event);
-});
-
-test('workflowError should not call job complete if the job is not active', async (t) => {
-  const state = {
-    reasons: {},
-    dataclips: {},
-    lastDataclipId: 'x',
-    activeJob: undefined,
-    activeRun: undefined,
-    errors: {},
-  };
-
-  const channel = mockChannel({
-    [RUN_COMPLETE]: (evt) => {
-      t.fail('should not call!');
-      return true;
-    },
-    [ATTEMPT_COMPLETE]: () => true,
-  });
-
-  const event = {
-    error: {
-      severity: 'crash',
-      type: 'Err',
-      message: 'it crashed',
-    },
-    state: {},
-  };
-
-  const context = {
-    channel,
-    state,
-    onFinish: () => {
-      t.pass();
-    },
-  };
-
-  await onWorkflowError(context, event);
-});
 
 // TODO what if an error?
 test('loadDataclip should fetch a dataclip', async (t) => {
